use global pref daysInMonth for AMC
Source: https://github.com/msupply-foundation/open-msupply/commit/7c9374f74f15336445d8015d8aa963450a6c23b3

+++ server/graphql/types/src/types/preferences.rs
@@ -63,7 +63,7 @@ impl PreferencesNode {
         self.load_preference(&self.preferences.adjust_for_number_of_days_out_of_stock)
     }
 
-    pub async fn days_in_month(&self) -> Result<i32> {
+    pub async fn days_in_month(&self) -> Result<f64> {
         self.load_preference(&self.preferences.days_in_month)
     }
